now n devices can play

+++ SampleInterface.java
@@ -23,7 +23,7 @@ public interface SampleInterface {
     public void music_data(byte[] data) throws BusException;
     
     @BusSignal
-    public void clock_sync(int countdown)throws BusException;
+    public void clock_sync(long countdown)throws BusException;
     
 }
